requirements.txt
openai==0.27.8
discord.py==2.3.1